requirements.txt
conllu
ipython
requests[socks]
webvtt-py
